Mypy typing fixes in several files

+++ sssnake/env/utils/schema.py
@@ -1,5 +1,8 @@
+from __future__ import annotations
+
 from typing import Callable, Dict, Sequence
 
+import numpy as np
 from gymnasium import spaces
 
 from sssnake.env.utils.config_def import EnvSpec
@@ -8,38 +11,38 @@ SpaceFactory = Callable[[EnvSpec], spaces.Space]
 
 OBS_SPACE_FACTORIES: Dict[str, SpaceFactory] = {
     "head_position": lambda spec: spaces.Box(
-        low=0.0, high=spec.max_map_size, shape=(2,), dtype=float
+        low=0.0, high=spec.max_map_size, shape=(2,), dtype=np.float32
     ),
-    "head_direction": lambda _spec: spaces.Box(low=0.0, high=360.0, shape=(), dtype=float),
+    "head_direction": lambda _spec: spaces.Box(low=0.0, high=360.0, shape=(), dtype=np.float32),
     "candy_position": lambda spec: spaces.Box(
-        low=0.0, high=spec.max_map_size, shape=(2,), dtype=float
+        low=0.0, high=spec.max_map_size, shape=(2,), dtype=np.float32
     ),
     "segments_num": lambda spec: spaces.Box(
         low=0,
         high=spec.tail_max_segment,
         shape=(),
-        dtype=int,
+        dtype=np.int64,
     ),
     "speed": lambda spec: spaces.Box(
-        low=spec.min_speed, high=spec.max_speed, shape=(), dtype=float
+        low=spec.min_speed, high=spec.max_speed, shape=(), dtype=np.float32
     ),
     "turnspeed": lambda spec: spaces.Box(
-        low=spec.min_turnspeed, high=spec.max_turnspeed, shape=(), dtype=float
+        low=spec.min_turnspeed, high=spec.max_turnspeed, shape=(), dtype=np.float32
     ),
     "map_size": lambda spec: spaces.Box(
-        low=spec.min_map_size, high=spec.max_map_size, shape=(), dtype=float
+        low=spec.min_map_size, high=spec.max_map_size, shape=(), dtype=np.float32
     ),
     "segments_positions": lambda spec: spaces.Box(
         low=0.0,
         high=spec.max_map_size,
         shape=(spec.tail_max_segment, 2),
-        dtype=float,
+        dtype=np.float32,
     ),
     "safe_map_snake": lambda spec: spaces.Box(
         low=0,
         high=1,
         shape=(spec.collision_map_resolution, spec.collision_map_resolution),
-        dtype=int,
+        dtype=np.int64,
     ),
 }
